+++ src/form/cqrd.jl
@@ -3,7 +3,7 @@
 
 
 function constraint_pipe_head_loss(
-    wm::CQRDWaterModel, n::Int, a::Int, node_fr::Int, node_to::Int, alpha::Float64,
+    wm::CQRDWaterModel, n::Int, a::Int, node_fr::Int, node_to::Int, exponent::Float64,
     L::Float64, r::Float64, q_max_reverse::Float64, q_min_forward::Float64)
     # Get the number of breakpoints for the pipe.
     num_breakpoints = get(wm.ext, :pipe_breakpoints, 1)