+++ test/runtests.jl
@@ -5,6 +5,7 @@ import JuMP
 import JSON
 import Ipopt
 import Juniper
+import Logging
 import Memento
 
 const _IM = WaterModels._IM
@@ -15,18 +16,13 @@ Memento.setlevel!(Memento.getlogger(_IM), "error")
 Memento.setlevel!(Memento.getlogger(Ipopt), "error")
 Memento.setlevel!(Memento.getlogger(Juniper), "error")
 WaterModels.logger_config!("error")
+Logging.disable_logging(Logging.Info)
 
 using Test
 
 # Default MIP and NLP optimizers.
 cbc = JuMP.optimizer_with_attributes(Cbc.Optimizer, "logLevel"=>0)
-
-ipopt = JuMP.optimizer_with_attributes(Ipopt.Optimizer, "tol"=>1.0e-8,
-    "acceptable_tol"=>1.0e-8, "max_iter"=>9999, "print_level"=>0, "sb"=>"yes")
-
-ipopt_ws = JuMP.optimizer_with_attributes(Ipopt.Optimizer, "tol"=>1.0e-8,
-    "acceptable_tol"=>1.0e-8, "max_iter"=>9999, "mu_init"=>1.0e-8,
-    "start_with_resto"=>"yes", "print_level"=>0, "sb"=>"yes")
+ipopt = JuMP.optimizer_with_attributes(Ipopt.Optimizer, "print_level"=>0, "sb"=>"yes")
 
 include("common.jl")